scraper/requirements.txt
supabase==2.10.0
beautifulsoup4==4.12.3
playwright==1.44.0
requests==2.32.3
python-dotenv==1.0.1
lxml==5.2.2
